+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- nested form elements to Freshdesk ticket creation form event
+
 ## [2.6.0] - 2023-05-12
 
 ### Added